DynamoDBStreams Now Returns Valid ShardIDs.
ShardIDs returned from the DynamoDBStreams emulator now mimic ShardIDs
in the production service.  Namely, they comply with the length requirements
in botocore.  Each returned ShardID is > 28 characters in length.

As ShardIDs from Kinesis are appended with `-$UUID5`.  The UUID5 value
generated is relative to the Kinesis stream name as well as the Kinesis
ShardID, ensuring that the value is consistent across multiple calls to
`describe_stream`.  Before ShardID values are passed through to
Kinesis, they are unwrapped to the original Kinesis value.