Instead of trying to set the default content-type on GET, set it on PUT
According to https://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectPUT.html
, if there is a PUT without a content-type, 'binary/octet-stream' will be used
as the default.

This changes the previous attempt at fixing this default, which was just setting
the content-type for GET's, which would override the content type of the
uploaded file, rather then respect the original uploaded content-type, like it
should have.