Add Filter Policy Support To Sns Subscriptions
See: https://docs.aws.amazon.com/sns/latest/dg/message-filtering.html

This commit adds filter policy support for SNS subscriptions. Filter
policies are set on the subscription and help filter the messages that
that subscription is interested in. If a message does not contain the
appropriate message attributes to pass the filter check, then the
message is not relayed to the subscriber. If the message does pass the
filter check then the message is passed to the subscriber.