Lambda: add function properties to LambdaContext for invocations
This adds some common properties to the LambdaContext, which is passed
to functions during invocations. In particular, the following properties
are added: function_name, function_version, invoked_function_arn.

The above properties are passed to lambda functions running inside docker.

Support for publishing during function creation has also been added,
which was a missing functionality, and was helpful in testing.

Contributed by @quantcast