Fix corner-case with `LOCALSTACK_` prefix strip.
PR 1181 introduced the ability to prefix localstack's environment
variables with `LOCALSTACK_` in the Docker image. However, it also
introduced a crash when a variable called `LOCALSTACK_` immediately
followed by a digit was present in the environment. The entrypoint
would try to re-export a variable starting with a digit which would
obviously fail.

The fix consist in skipping problematic variables in the entrypoint.
Such variable don't make sense for localstack anyways.

Fixes #1387