Use the correct region when getting CloudFormation stack resources to delete.
CloudFormation's DeleteStack wrapper would create an `aws_stack` client in
order to list the resources contained in the stack for deletion. This
client was created without specifying a `region_name` parameter which meant
that even if the stack was created in a given region the deletion would
always be attempted in us-east-1.

This change fixes that by specifying the region name sent in the request
itself.