Save e-mail contents in easy, machine-readable format
When using loclstack for development an even for easier integration/local testing
of interaction with AWS services, a feature is missing to check if an e-mail
was sent correctly.

Issue #2514 was closed with a patch that dumped the e-mail contents to the
log output. However, in most cases, this is not reliably useable for testing,
as it would require to parse the log-stream of localstack.

This commit now adds a feature where a sent e-mail is persisted to the local
file system. It re-uses the specified data directory (DATA_DIR) and falls back
to the temp directory to save e-mails into an `ses` subdirectory. Each
successfully sent message is saved in a JSON file named after the message ID.
The contents are the Source, Subject, Body, Destinations (To, CCs, BCCs) as
well as the region.

Co-Authored-By: Arne Wohlert <arne.wohlert@oss.volkswagen.com>