Similar to https://github.com/localstack/localstack/pull/4098 I needed to check if an email was sent correctly but relied upon templated emails which the above PR did not address.
I have essentially duplicated the work done in the above PR but amended it slightly
for the send_templated_email use case i.e. sending a templated email will dump said
email into the DATA_DIR in an ses subdirectory using the id of the sent email as
filename.

The contents of the JSON file are:
- Source
- Template
- TemplateData
- Destinations

e.g.

{
  "Source": "user@example.com",
  "Template": ["hello-world"],
  "TemplateData": ["{\"A key\": \"A value\"}"],
  "Destinations": {
    "ToAddresses": ["success@example.com"],
    "CcAddresses": [],
    "BccAddresses": []
  },
  "Region": "us-east-1"
}