add attribute md5 generation
The md5 method was copied and adapated from moto. Alternatively,
you can also wrap the message accordingly for moto and pass
it. Such a method exists in sqs_listener.py, however
depending on the old listener seems conceptually wrong. It
is also flagged as probably unnecessary. So all approaches seemed
problematic to a degree.