Changed postional vars to named
The `find_change_set` in the`get_template` function had its parameters mistakenly inverted. This caused a GetTemplate request which had both a stack-name and change-set-name to fail.

I changed the passed parameters to named parameters because I'm unsure if there are other areas of localstack which also call `find_change_set` and didn't want to risk breaking them.