feat: implement sns -> ses event emitting
This test sets up a verified email, connects an SNS topic to SES and sends an
email. To test this, the test connects an SQS queue to the SNS topic to
ensure the messages have been received correctly.

We use the new `sqs_receive_num_messages` fixture from the previous
commit to receive the required three messages:

1. SES sends a test message to the topic when the configuration set is
   connected
2. SES sends a "sent" event
3. SES sends a "delivered" event

See #7184