Make /health support HEAD query again
HEAD query support was disabled as part of the refactoring in #4883.

It used to be supported by the following implementation.

```python
def handle(self, method, path, data) -> Optional[Dict]:
    if method == "GET":
        return self.get(path, data)
    if method == "POST":
        return self.post(path, data)
    if method == "PUT":
        return self.put(path, data)

    return {}
```

When using LocalStack in a Docker environment, third-party tools like
`wait-for-200` might be used to probe the readiness of the LocalStack
service. And some of these tools use HEAD queries by default. In such
environments, upgrading from an older version of LocalStack might break
some pods accidentally and it's not straightforward to debug because you
can only see the following output in your LocalStack logs:

```
2022-11-29T09:03:01.545 ERROR --- [   asgi_gw_1] l.aws.handlers.logging
: exception during call chain: 405 Method Not Allowed: The method is not
allowed for the requested URL.
```

This PR attempts to add HEAD query support back to ease the upgrade for
users with such setups.