prevent DEFAULT_REGION's deprecation message to always show
Whenever we start localstack we're greeting with this warning message:

  WARN --- [  MainThread] localstack.deprecations    : DEFAULT_REGION is deprecated (since 0.12.7) and will be removed in upcoming releases of LocalStack! LocalStack now has full multi-region support. Please remove this environment variable.

This happens even if we don't have the environment variable DEFAULT_REGION
set. This is because in `localstack/config.py` the check for the
environment variable has a side effect of creating it if the environment
variable is not set!

Later on the detection `log_deprecation_warnings()` checks kick-in and
detect that the environment variable is set and issue the warning.

With this commit we're avoiding creating the environment variable.