Support custom key material when creating KMS keys
Add support for creating KMS keys with custom key material using custom
tags. This is similar to an existing feature that allows the KMS Key ID
to be set when creating KMS keys. This is necessary when working with
external systems and decryption or HMACs must work across localstack
reboots.

Fixes #8910