Respect file extension setting for S3 destinations in Firehose provider
The Firehose delivery stream implementation was not considering the file
extension setting when the target was defined as an S3 bucket. This fix
ensures that the specified file extension is appended to the S3 object's
key when writing data. The `_get_s3_object_path` method now accepts the
`file_extension` parameter to accommodate this enhancement.