add IPv6 support to the localstack runtime gateway
Adds support for configuring the gateway listener to listen on IPv6
addresses. For example, you can now set:

```
GATEWAY_LISTEN="[::]"
```

NOTE: `GATEWAY_LISTEN` now expects IPv6 address to be enclosed in square
brackets, since the existing interface expected for `GATEWAY_LISTEN` is
a subset of a URL (host and port), and square brackets are required for
IPv6 addresses in URLs (to distinguish between the colons in an IPv6
address and the delimiter with the port number). This format is defined
in https://www.rfc-editor.org/rfc/rfc6874.

This required modifications to:

* The HostAndPort.parse() method, which can now interpret an IPv6
    address correctly.
* The UniqueHostAndPortList class, which now distinguishes between IPv6
    and IPv4 addresses when deduping.
* The Twisted gateway, which will now use an IPv6 endpoint for an IPv6
    host.

I didn't make any changes to the two other gateway implementations
(hypercorn or werkzeug) as I'm not sure if they're still considered
supported.

I also noticed a bug in the existing twisted implementation, in which it
ignores any IPv4 host passed in via GATEWAY_LISTEN, and simply binds to
all interfaces ("", which is interpreted as 0.0.0.0). I didn't fix this
(given users might be depending on the current behavior), but the new
IPv6 implementation respects the specified host.

Addresses https://github.com/localstack/localstack/issues/11600