Add composite alarms store
Leaving current `alarms` property intact for metric alarms intact, without renaming. Most AWS commands like describe-alarms only return metric alarms if composite alarm type is not specified explicitly as a parameter. Therefore it is reasonable to keep the store name as is and add a separate store for composite ones.