Skip rule evaluation on missing alarms
As discussed in the (PR)[https://github.com/localstack/localstack/pull/11828#discussion_r1838441628], partial behaviour should not hurt parity. In this case, because evaluation by name is not yet implemented, it is better to let user save an alarm that is not fully supported by Localstack but skip this alarm evaluation.