Match an empty prefix or suffix eventbridge pattern
I was using this pattern to match all events:
https://stackoverflow.com/questions/62406933/aws-eventbridge-pattern-to-capture-all-events

It works in AWS and in localstack before 4.0.3.

Because the new rule engine was using truthiness to determine if the
condition was a prefix one in `condition.get('prefix')`, it would assume
a prefix pattern of `''` isn't actually a prefix pattern. This instead
checks for `None` to determine the condition type and gets empty
prefixes to match again.