Fix passing of tags into moto on subnet create
This fixes a recent regression whereby the tags dict was inadvertently altered as part of processing to determine if a custom id was specified.  When passed to the moto library, the tags were not in the expected format and ignored.  This fix preserves the original format and adds a test case.