Fix subnet tags going missing (#12459)
This fixes a recent regression whereby the tags dict was inadvertently altered as part of processing to determine if a custom id was specified.  When passed to the moto library, the tags were not in the expected format and ignored.  This fix preserves the original format and adds a test case.

Tags do not form part of the SG, VPC and subnet objects in moto.  Instead they are stored in a separate dict in the EC2 backend.  This means that if the object id is being changed, the tags need to have their key updated in the tags dict.

Added cleanup processing to subnet and SG test cases.