Use same util function to get current time
datetime.utcnow() is deprecated and was giving a 2-hour difference result than get_current_time() for a computer running in CEST timezone.
This led to valid records being expired in bisect_events_by_record_age in stream poller.

Same function needs to be used across the code, especially when a comparison is made within same logic.