Add mock result validation according to field validation modes
TestState validates mocked responses against AWS API models with three validation modes: STRICT (this is the default and validates all required fields), PRESENT (validates field types and names), and NONE (no validation).

STRICT field validation mode checks the presence of required fields and validates the type and format of the fields in the mock against the respective service shape in AWS API definitions.
PRESENT field validation mode doesn't check the presence of required fields. It only validates the type and format of the fields that are present in the user-provided mock.
NONE mode means no format validation will be performed. A check that the result is a valid JSON string is still performed even in NONE field validation mode.

Note: fields that are not present in the API spec but are supplied in the mock will be allowed, because of forward-compatibility.