feat(sfn/test_state): Support mocking ContextObject (#13431)
Adds limited support for mocking a ContextObject in the SFN TestState API.

Empty context case is not accounted for.

Changes
Uses pydantic (and some python magic) to validate that passed in context, and bubble up a validationexception.
Reworks the inspectionData approach to instead patch each _eval_body and add appended results for a given key