dev: support executing kubernetes dev setup on command execution
# Motivation

When using our kubernetes dev command to generate the config/override files for a k3d cluster, it dumps the files to disk and prints the commands needed to the terminal. The next step for the user is then to copy and paste those commands from their terminal and wait for them.

Wouldn't it be nice to have a single command that deploys the cluster and LocalStack at the same time?

# Changes

* add an `-x/--execute` flag to the CLI (which implies the `--write` flag) which executes the commands